Migrating datastore to new CI.
