+++ datastore/tests/system/test_system.py
@@ -14,8 +14,6 @@
 
 import datetime
 import os
-import pkgutil
-import tempfile
 import unittest
 
 import httplib2
@@ -27,27 +25,11 @@ from google.cloud.datastore.helpers import GeoPoint
 from google.cloud.environment_vars import GCD_DATASET
 from google.cloud.exceptions import Conflict
 
-import clear_datastore
-import populate_datastore
-from system_test_utils import EmulatorCreds
-from system_test_utils import unique_resource_id
+from test_utils.system import EmulatorCreds
+from test_utils.system import unique_resource_id
 
-
-SPHINX_CONF = """\
-extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.doctest',
-]
-"""
-
-SPHINX_SECTION_TEMPLATE = """\
-Section %02d
-===========
-
-.. automodule:: google.cloud.%s
-  :members:
-
-"""
+from tests.system.utils import clear_datastore
+from tests.system.utils import populate_datastore
 
 
 class Config(object):
@@ -521,56 +503,3 @@ class TestDatastoreTransaction(TestDatastore):
                 # transaction.
                 entity_in_txn[contention_prop_name] = u'inside'
                 txn.put(entity_in_txn)
-
-
-class TestDoctest(unittest.TestCase):
-
-    def _submodules(self):
-        pkg_iter = pkgutil.iter_modules(datastore.__path__)
-        result = []
-        for _, mod_name, ispkg in pkg_iter:
-            self.assertFalse(ispkg)
-            result.append(mod_name)
-
-        self.assertNotIn('__init__', result)
-        return result
-
-    @staticmethod
-    def _add_section(index, mod_name, file_obj):
-        mod_part = 'datastore'
-        if mod_name != '__init__':
-            mod_part += '.' + mod_name
-        content = SPHINX_SECTION_TEMPLATE % (index, mod_part)
-        file_obj.write(content)
-
-    def _make_temp_docs(self):
-        docs_dir = tempfile.mkdtemp(prefix='datastore-')
-
-        conf_file = os.path.join(docs_dir, 'conf.py')
-
-        with open(conf_file, 'w') as file_obj:
-            file_obj.write(SPHINX_CONF)
-
-        index_file = os.path.join(docs_dir, 'contents.rst')
-        datastore_modules = self._submodules()
-        with open(index_file, 'w') as file_obj:
-            self._add_section(0, '__init__', file_obj)
-            for index, datastore_module in enumerate(datastore_modules):
-                self._add_section(index + 1, datastore_module, file_obj)
-
-        return docs_dir
-
-    def test_it(self):
-        from sphinx import application
-
-        docs_dir = self._make_temp_docs()
-        outdir = os.path.join(docs_dir, 'doctest', 'out')
-        doctreedir = os.path.join(docs_dir, 'doctest', 'doctrees')
-
-        app = application.Sphinx(
-            srcdir=docs_dir, confdir=docs_dir,
-            outdir=outdir, doctreedir=doctreedir,
-            buildername='doctest', warningiserror=True, parallel=1)
-
-        app.build()
-        self.assertEqual(app.statuscode, 0)
