+++ datastore/tests/unit/test_batch.py
