+++ datastore/tests/unit/test_client.py
