--- datastore/tox.ini
@@ -1,35 +1,0 @@
-[tox]
-envlist =
-    py27,py34,py35,cover
-
-[testing]
-localdeps =
-    pip install --quiet --upgrade {toxinidir}/../core
-deps =
-    {toxinidir}/../core
-    mock
-    pytest
-covercmd =
-    py.test --quiet \
-      --cov=google.cloud.datastore \
-      --cov=unit_tests \
-      --cov-config {toxinidir}/.coveragerc \
-      unit_tests
-
-[testenv]
-commands =
-    {[testing]localdeps}
-    py.test --quiet {posargs} unit_tests
-deps =
-    {[testing]deps}
-
-[testenv:cover]
-basepython =
-    python2.7
-commands =
-    {[testing]localdeps}
-    {[testing]covercmd}
-deps =
-    {[testenv]deps}
-    coverage
-    pytest-cov
